:telephone_receiver: field_deleteone() now delete only if field is all NULL

+++ data/field.sql
@@ -43,14 +43,17 @@ $$ LANGUAGE plpgsql;
 CREATE OR REPLACE FUNCTION "field_deleteone" (_a JSONB)
 RETURNS VOID AS $$
 BEGIN
-  -- 1. delete from table and drop column
-  DELETE FROM "field" WHERE "id"=_a->>'id';
-  EXECUTE format('ALTER TABLE "food" DROP COLUMN IF EXISTS %I', _a->>'id');
+  -- 1. delete from table and drop column, only if empty
+  EXECUTE format('SELECT * FROM "food" WHERE %I IS NOT NULL', _a->>'id');
+  IF NOT FOUND THEN
+    DELETE FROM "field" WHERE "id"=_a->>'id';
+    EXECUTE format('ALTER TABLE "food" DROP COLUMN IF EXISTS %I', _a->>'id');
+  END IF;
 END;
 $$ LANGUAGE plpgsql;
 
 
-CREATE OR REPLACE FUNCTION "type_upsertone" (_a JSONB)
+CREATE OR REPLACE FUNCTION "field_upsertone" (_a JSONB)
 RETURNS VOID AS $$
 BEGIN
   PERFORM type_deleteone(_a);
